fork tsan check

+++ src/threadpool.cpp
@@ -39,7 +39,7 @@ void Threadpool::work(){
 
 		//try to steal from other random queues
 		std::uniform_int_distribution<int> distribution(0,worker_id_ticket.load() -1);
-		for(int i=0; i<worker_id_ticket.load()-1; i++){
+		for(size_t i=0; i<worker_id_ticket.load()-1; i++){
 			int random_index = distribution(random_generator);
 		
 			if(queues.at(random_index).try_steal(task)){
